+++ lib/interface-types/src/interpreter/instructions/mod.rs
@@ -1,14 +1,14 @@
 mod argument_get;
 mod call_core;
 mod lowering_lifting;
-mod read_utf8;
-mod write_utf8;
+mod memory_to_string;
+mod string_to_memory;
 
 pub(crate) use argument_get::argument_get;
 pub(crate) use call_core::call_core;
 pub(crate) use lowering_lifting::*;
-pub(crate) use read_utf8::read_utf8;
-pub(crate) use write_utf8::write_utf8;
+pub(crate) use memory_to_string::memory_to_string;
+pub(crate) use string_to_memory::string_to_memory;
 
 #[cfg(test)]
 pub(crate) mod tests {
@@ -131,23 +131,12 @@ pub(crate) mod tests {
                             },
                         },
                     );
-                    hashmap.insert(
-                        "alloc".into(),
-                        Export {
-                            inputs: vec![InterfaceType::I32],
-                            outputs: vec![InterfaceType::I32],
-                            function: |arguments: &[InterfaceValue]| {
-                                let _size: i32 = (&arguments[0]).try_into().unwrap();
-
-                                Ok(vec![InterfaceValue::I32(0)])
-                            },
-                        },
-                    );
 
                     hashmap
                 },
                 locals_or_imports: {
                     let mut hashmap = HashMap::new();
+                    // sum
                     hashmap.insert(
                         42,
                         LocalImport {
@@ -158,6 +147,19 @@ pub(crate) mod tests {
                                 let b: i32 = (&arguments[1]).try_into().unwrap();
 
                                 Ok(vec![InterfaceValue::I32(a * b)])
+                            },
+                        },
+                    );
+                    // string allocator
+                    hashmap.insert(
+                        43,
+                        LocalImport {
+                            inputs: vec![InterfaceType::I32],
+                            outputs: vec![InterfaceType::I32],
+                            function: |arguments: &[InterfaceValue]| {
+                                let _size: i32 = (&arguments[0]).try_into().unwrap();
+
+                                Ok(vec![InterfaceValue::I32(0)])
                             },
                         },
                     );